Switch type to float64

+++ api/types.go
@@ -45,7 +45,7 @@ type AlertConditionTerm struct {
 	Duration     int     `json:"duration,string,omitempty"`
 	Operator     string  `json:"operator,omitempty"`
 	Priority     string  `json:"priority,omitempty"`
-	Threshold    float32 `json:"threshold,string,omitempty"`
+	Threshold    float64 `json:"threshold,string,omitempty"`
 	TimeFunction string  `json:"time_function,omitempty"`
 }
